bug 1468987: kibana_proxy OOM
We currently set the memory allocated to the kibana-proxy container to be
the same as `max_old_space_size` for nodejs. But in V8, the heap consists
of multiple spaces.

The old space has only memory ready to be GC and measuring the used heap
by kibana-proxy code, there is at least additional 32MB needed in the code
space when `max_old_space_size` peaks.

Increasing the default memory limit to 256MB in openshift-ansible repository
and also changing the default calculation of `max_old_space_size` here to be
only half of what the container receives to allow some memory for other heap
spaces.